Moved Home page molecules to a folder in the Molecule directory to clean up the structure even more.
