+++ src/components/Organisms/HomePage/HomepageContainer.js
@@ -1,8 +1,8 @@
 import "./homepageContainer.css";
-import HomeHeader from "./Header/Header";
-import Testimonials from "./TestimonialSection/TestimonialsSection";
-import Services from "./ServicesSection/ServicesSection";
-import Contact from "./ContactSection/ContactSection";
+import HomeHeader from "../../Molecules/HomeMolecules/Header/Header";
+import Testimonials from "../../Molecules/HomeMolecules/TestimonialSection/TestimonialsSection";
+import Services from "../../Molecules/HomeMolecules/ServicesSection/ServicesSection";
+import Contact from "../../Molecules/HomeMolecules/ContactSection/ContactSection";
 import NavBar from "../../Molecules/NavBar/NavBar";
 import Footer from "../..//Molecules/Footer/Footer";
